feat(utils): abortable wait
Add support for `AbortController` in `wait` which allows for premature
cancellation of wait.
