+++ packages/utils/src/exports.ts
@@ -1,3 +1,4 @@
+import { AbortError } from './AbortError'
 import { Defer } from './Defer'
 import { ENSName, toENSName } from './ENSName'
 import { EthereumAddress, toEthereumAddress } from './EthereumAddress'
@@ -21,7 +22,7 @@ import { toEthereumAddressOrENSName } from './toEthereumAddressOrENSName'
 import { BrandedString } from './types'
 import { wait } from './wait'
 import { waitForEvent } from './waitForEvent'
-import { AbortError, TimeoutError, withTimeout } from './withTimeout'
+import { TimeoutError, withTimeout } from './withTimeout'
 
 export {
     BrandedString,
