+++ packages/utils/test/wait.test.ts
@@ -1,13 +1,28 @@
 import { wait } from '../src/wait'
+import { AbortError } from '../src/AbortError'
 
 describe(wait, () => {
     // https://stackoverflow.com/questions/21097421/what-is-the-reason-javascript-settimeout-is-so-inaccurate
     const JITTER_FACTOR = 4
 
-    it("waits at least the predetermined time", async () => {
+    it('waits at least the predetermined time', async () => {
         const start = Date.now()
         await wait(20)
         const end = Date.now()
         expect(end - start).toBeGreaterThanOrEqual(20 - JITTER_FACTOR)
+    })
+
+    it('rejects if aborted during wait', () => {
+        const abortController = new AbortController()
+        setTimeout(() => {
+            abortController.abort()
+        }, 10)
+        return expect(wait(20, abortController)).rejects.toEqual(new AbortError())
+    })
+
+    it('rejects if initially aborted', () => {
+        const abortController = new AbortController()
+        abortController.abort()
+        return expect(wait(20, abortController)).rejects.toEqual(new AbortError())
     })
 })
