+++ packages/utils/test/withTimeout.test.ts
@@ -1,4 +1,5 @@
-import { AbortError, TimeoutError, withTimeout } from '../src/withTimeout'
+import { TimeoutError, withTimeout } from '../src/withTimeout'
+import { AbortError } from '../src/AbortError'
 
 describe(withTimeout, () => {
     it('resolves if given promise resolves before timeout', () => {
